IMS-VT: Fix crash due to SurfaceTexture released wrongly in VT call.
Whenever there is a change in the primary video call,
VideoCallFragment is re created for new call. Old
VideoCallFragment instance is destroyed and new one is created.
As part of Fragment destroy, the onSurfaceTextureDestroyed() call
back is invoked on the listener. Since we have to always pass
the same SurfaceTexture to ims media till all calls are ended,
we return false in onSurfaceTextureDestroyed callback to
prevent SurfaceTexture from being released.
There is a race condition scenario where TextureViews of
already destroyed VideoCallFragment is attached to the
SurfaceTexture during VideoCallFragment re create. This
causes the listeners for new VideoCallFragment TextureViews
to be cleared. When the new VideoCallFragment is later
destroyed, there is no onSurfaceTextureDestroyed() callback
invoked and the SurfaceTexture is released. Later when this
released SurfaceTexture is used, the application crashes.
To fix this we need to prevent old VideoCallFragment from attaching
the views to SurfaceTexture. This is done by setting the
videoCallSreen to null in VideoCallPresenter
onVideoCallScreenUiUnready() function. This will make
sure that calls to showVideoViews and other UI related
operations are not performed when VideoCallFragment
is destroyed or is in background.

Change-Id: I6f426350b4201c01c9c548e2bd33740a74a1dc92
CRs-Fixed: 2671948