Fix Dialer NPE during call
After call end, and it is cleared up, hence the context of InCallPresenter
is set to null, once the InCallActivity is not finished yet, at this
time, receive intent to restart InCallActivity, later call onPause
as per its lifecycle to perform onUiShowing, then NPE happens.

To fix it, add null check.

Change-Id: If8f66b76928101e2836ff5b54a0a5d695125ad6e
CRs-Fixed: 2168836