Fix the prompt window is still there during emergency call
While APM is on, make a call to show a prompt dialog, if you don't
dismiss the dialog, then click home key to go to dialer and make an
emergency call, at that time, the prompt dialog is still shown.

To fix it, dismiss the dialog while new MO call is starting.

Change-Id: Ia76186f4921020178cd36664abd641d6d229d12e
CRs-Fixed: 2208957