Close the pending dialog properly
If InCallActivity with pending dialog goes to background, it won't
be destroyed. while the screen is on, it is resumed so that it is
re-created for new call, which is not needed. Otherwise, in case of
re-creation, the dialog also is leaked.

To fix it, close the pending dialog in method "onStop".

Change-Id: I7c7feacf84e0454bdbfae8fee6c05bb05c5bc58f
CRs-Fixed: 2210097