Fix In-call UI is stuck in selecting phone account
If user quickly triggers 2 MO calls in short time, the phone account
selecting dialog just shows for the fisrt call only, but the first call
will be disconnected by Telecom, hence, the selection is ignored due
to null call instance. Meantime the second call is pending on phone
account selection, but UI does not start the dialog due to the same call
state.

To fix it, allow the phone account selection dialog to multi-MO call.

Change-Id: Ie77c7c2fe91454ec254e2e53760b45b41cd1a64a
CRs-Fixed: 2311876