Fix wrong entering search UI
Slide-in animation is used to show smart search UI after dialpad
is shown, but the animation might not be finished after dialpad is hidded
while quickly clicking dial icon and dismissing phone account selection
dialog to go back to the favorite table.

To fix it, don't request to enter the search UI if dialpad is hidded.

Change-Id: I2783e14226e9632fe312fe32690ab2afa67e8af4
CRs-Fixed: 2311124