IMS-VT: Update secondary info when upgrade operation is done
When MO upgrade is initiated, CallCardPresenter is informing
its clients not to show secondary info but is not informing
back to its clients to show secondary info after the upgrade
operation is completed. Now, when user exits full screen mode 
in video call screen, buttons will be refreshed. 
As part of this, swap icon and hold banner are not shown 
since CallCardPresenter informed its clients not to show
secondary info.

Fix this in CallCardPresenter to inform its clients to
show secondary info after the upgrade operation is
completed.

Change-Id: Ibea5aeaa4fef151db844a8d17d555450d44180b5
CRs-Fixed: 2478539